+++ src/EightFaceMk2Base.hpp
@@ -45,7 +45,7 @@ struct EightFaceMk2Base : Module, StripIdFixModule {
 
 	LongPressButton presetButton[NUM_PRESETS];
 
-	int ctrlModuleId = -1;
+	int64_t ctrlModuleId = -1;
 	int ctrlOffset = 0;
 	/** Current operating mode */
 	CTRLMODE ctrlMode = CTRLMODE::READ;
